requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# AI/ML
google-generativeai==0.8.3
PyPDF2==3.0.1

# Caching & Performance
redis==5.0.1
python-jose[cryptography]==3.3.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0